Added ASIO2WASAPI plugin so low latency exclusive mode WASAPI is available for everyone on Vista+
Added ASIO2WASAPI plugin so low latency exclusive mode WASAPI is available for everyone on Vista+.
Contrary to original ASIO2WASAPI driver this one is a private plugin so it is not registered in the system and so can co-exists with the original one.
Although if you know you do not need this plugin it can be de-selected in the installer.
Be aware that since this is an exclusive mode plugin it uses the audio hardware exclusively so other apps cannot use it at the same time. So its usage is only recommended when low latency is a requirement.